Few changes to SchoolDistricts and Users table

+++ src/main/java/com/mealManage/repository/SchoolRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.mealManage.model.School;
 
 @RepositoryRestResource(collectionResourceRel = "school", path = "school")
-public interface SchoolRepository extends PagingAndSortingRepository<School, Long>{
+public interface SchoolRepository extends PagingAndSortingRepository<School, Integer>{
 	List<School> findBySchoolDistrictId(@Param("schoolDistrictId") Integer schoolDistrictId);
 }
